% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc_ci.R
\name{cuminc_ci}
\alias{cuminc_ci}
\title{Compute confidence intervals for cumulative incidence}
\usage{
cuminc_ci(ci_fit, times, conf_level = 0.95)
}
\arguments{
\item{ci_fit}{The object returned by}

\item{times}{A vector of times at which to estimate the confidence interval}

\item{conf_level}{The confidence interval; default is 0.95.}
}
\value{

}
\description{
Compute confidence intervals for cumulative incidence
}
\examples{
# See Tringale-Second Cancer project
}
