% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genie_investigate_pt.R
\name{genie_investigate_pt}
\alias{genie_investigate_pt}
\title{Investigate GENIE BPC Patient}
\usage{
genie_investigate_pt(pt)
}
\arguments{
\item{pt}{record_id to investigate}
}
\value{

}
\description{
Investigate GENIE BPC Patient
}
