% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msk_colors_rgb.R
\docType{data}
\name{msk_colors_rgb}
\alias{msk_colors_rgb}
\title{Named list of MSK branded colors in RGB format}
\format{
An object of class \code{character} of length 30.
}
\usage{
msk_colors_rgb
}
\description{
This is named vector of the MSK branded color hex codes.
Run \code{msk_colors_rgb} to
see the names and RGB codes for the MSK branded colors.
}
\keyword{datasets}
