% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_function.R
\name{survival_function}
\alias{survival_function}
\title{A function for running common pieces of a survival analyses: (1) the KM
model (2) the median survival (3) the KM curve}
\usage{
survival_function(
  df,
  time1,
  time,
  status,
  covar,
  time_units = "Months",
  time_origin = "Diagnosis"
)
}
\arguments{
\item{df}{The data frame containing time, status and covariate variables}

\item{time1}{If the analysis involves left truncation, this is the time
from the origin to entry into the risk set}

\item{time}{The time from the origin to the event or end of follow-up}

\item{covar}{The covariate of interest}

\item{time_units}{An optional character string indicating the units of time.
The default is "Months". This is used to properly label the units for median
survival and the x-axis on the survival plot.}

\item{time_origin}{An optional character string indicating the time origin.
The default is "Diagnosis". This is used to label the x-axis on the survival
plot.}
}
\value{

}
\description{
A function for running common pieces of a survival analyses: (1) the KM
model (2) the median survival (3) the KM curve
}
